+++ Explorer/Assets/Scripts/CrdtEcsBridge/Components/Defaults/PBMeshColliderDefaults.cs
@@ -13,13 +13,4 @@ namespace CrdtEcsBridge.Components.Defaults
         public static uint GetColliderLayer(this PBMeshCollider self) =>
             self.HasCollisionMask ? self.CollisionMask : (uint)ColliderLayer.ClPhysics | (uint)ColliderLayer.ClPointer;
     }
-
-    public static class PBMeshRendererDefaults
-    {
-        public static float GetTopRadius(this PBMeshRenderer.Types.CylinderMesh self) =>
-            self.HasRadiusTop ? self.RadiusTop : 0.5f;
-
-        public static float GetBottomRadius(this PBMeshRenderer.Types.CylinderMesh self) =>
-            self.HasRadiusBottom ? self.RadiusBottom : 0.5f;
-    }
 }
