+++ Explorer/Assets/Scripts/CrdtEcsBridge/Components/Defaults/PBMeshRendererDefaults.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2c227b0b3477184880fb553963676d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
